moved restricted sale days related variables into struct
This is a WIP commit: the contracts compile, but the move into a struct seems to have broken TestQINTokenSale2.

+++ contracts/permissions/BuyerStore.sol
@@ -10,7 +10,7 @@ import '../permissions/Ownable.sol';
 */
 contract BuyerStore is Ownable {
     using SafeMath for uint8;
-    using SafeMath for uint256;
+    using SafeMath for uint;
 
     uint public registeredUserCount = 0;
